feat: apply profile phone number get logic (#381)
